+++ src/WeihanLi.Npoi/InternalConstants.cs
@@ -16,15 +16,5 @@
         /// ApplicationName
         /// </summary>
         public const string ApplicationName = "WeihanLi.Npoi";
-
-        /// <summary>
-        /// InputFormatterFuncName
-        /// </summary>
-        public const string InputFormatterFuncName = "InputFormatterFunc";
-
-        /// <summary>
-        /// OutputFormatterFunc
-        /// </summary>
-        public const string OutputFormatterFuncName = "OutputFormatterFunc";
     }
 }